init vote validation logic split
